# Improve write performance for operations on tables with indexes + Bugfix for `remove`