--- a/packages/db/src/core/createDB.ts
+++ b/packages/db/src/core/createDB.ts
@@ -23,7 +23,8 @@ export function createDB(options?: Partial<DBOptions>): Database {
 
 export interface DBOptions {
   /**
-   * Toggles whether entities are cloned before being returned from functions like `many()`, `first()` or `one()`.
+   * Toggles whether entities are cloned before being written with function `insert` or
+   * returned from functions like `many()`, `first()` or `one()`.
    *
    * If enabled, adds a performance cost, but prevents the user from modifying
    * the returned entities directly, which would bring the database into an inconsistent state.