@froehlichA 2022-09-13
Thanks for creating a new benchmark - I'll eventually remove these files altogether, and build some benchmarks right into the docs so I can

1. Test BlinkDB unter realistic circumstances (i.e. in the browser)
3. Compare it to Redux or similar store alternatives (right now I'm only comparing performance to lokijs in order to make sure I'm not writing poorly performing code)
2. Showcase the speed :P